refactor: rename subdomain step file
